backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
PyJWT==2.10.1

# AWS Services
boto3==1.35.8
botocore==1.35.8


# SSE
sse-starlette==1.8.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7

# PDF / Docs
pypdf==3.17.4
reportlab==4.0.9
pymupdf==1.24.10
pytesseract==0.3.10
pdf2image>=1.17.0
Pillow==10.4.0

# Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
#httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.50.0

# Dev Tools (Optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pymupdf==1.24.10

pdfplumber==0.11.4
jsonschema==4.23.0

cryptography>=42.0.0
httpx>=0.27.0
pyyaml>=6.0
apscheduler>=3.10.0
python-docx==1.1.2

# Payment gateway
razorpay==1.4.2

# Firecrawl — URL reading and web search
firecrawl-py>=1.9.0
